+++ lib/nostr_sdk.dart
@@ -13,3 +13,4 @@ export 'src/rust/api/protocol/signer.dart';
 export 'src/rust/api/relay/options.dart';
 export 'src/rust/api/client.dart';
 export 'src/rust/api/client/options.dart';
+export 'src/rust/api/client/output.dart';